get file resolution and duration on load

+++ gem_mpv.hpp
@@ -189,8 +189,8 @@ class GEM_EXTERN mpv : public gemframebuffer
     bool m_size_changed{};
     bool m_auto_resize{};
 
-    int m_media_width{512};
-    int m_media_height{512};
+    int64_t m_media_width{512};
+    int64_t m_media_height{512};
 
     t_outlet* m_prop_outlet{};
 };
